requirements.txt
folium==0.14.0
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
plotly==5.15.0
Requests==2.31.0
seaborn==0.12.2
streamlit==1.23.1
streamlit_folium==0.12.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
